continue refactoring cfn/index module into separate files
